(refactor) Move logic from useFileHref and Header handleFileUpload to useStrategyFile
